+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Added parsing support for PEP 448 unpacking generalizations.
 - Added parsing support for PEP 515 underscores in numeric literals.
 
+### Fixed
+- Fixed parsing variable with annotation only.
+
 ## 1.0.1
 
 ### Fixed