+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added parsing support for PEP 465 matrix multiplication operators.
 - Added parsing support for PEP 448 unpacking generalizations.
 - Added parsing support for PEP 515 underscores in numeric literals.
+- Added parsing support for PEP 570 positional only parameters.
 
 ### Fixed
 - Fixed parsing variable with annotation only.